requirements.txt
torch>=2.3.0
transformers>=4.51.0
datasets>=3.0.0
accelerate>=1.0.0
peft>=0.14.0
trl>=0.16.0
bitsandbytes>=0.45.0
sentencepiece>=0.2.0
sacrebleu>=2.4.0
evaluate>=0.4.3
gradio>=5.23.0
librosa>=0.10.2
soundfile>=0.12.1
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.5.0
pyyaml>=6.0.2
python-dotenv>=1.0.1
huggingface_hub>=0.30.0
llama-cpp-python>=0.3.0
